Rename networks to tokens. Rename mainnet to ethereum
