+++ src/tokenlists/balancer/tokens/ethereum.ts
