+++ src/tokenlists/balancer/tokens/goerli.ts
